requirements.txt
playwright
beautifulsoup4
schedule
python-dotenv
